Review help pages
Fix link in legal.xml
Review index.page and introduction.page: mark as final.

Shorten prob-error-while-retrieving-definition to
prob-retrieving-definition because it was repeating itself and it is
best to keep page names short. Update links to match the new page ID.

Similar Words shows other dictionaries in the source, not other sources.
Update the page so that it is more precise.