app: Remove Copy menu entry
The widgets used for the search entry and for showing the definition
already handle copying selected text. There is no need for an entry
menu to do this, plus we are a step closer to get rid of the Edit menu.

See https://bugzilla.gnome.org/show_bug.cgi?id=334866