gdict: Use a symbolic icon in DatabaseChooserButton
Using text is going to break it, especially when we're using the
internal database name from the dictionary context, which may or may not
be a meaningful, or even translatable string.

Additionally, the DICT protocol specifies special names that would need
to be converted internally to something that is user readable.

Instead of this insanity, let's just a simple symbolic icon for lists of
things — which is appropriate in the case of a list of databases.